+++ src/components/reusables/Input.tsx
@@ -8,6 +8,7 @@ export interface InputProps {
   callback?: (e: any) => any;
   value?: any;
   checked?: boolean;
+  id: string;
 }
 
 export default class Input extends React.Component<InputProps, {}> {
@@ -23,7 +24,7 @@ export default class Input extends React.Component<InputProps, {}> {
           ref="input" className="form-control"
           type={this.props.type || "text"}
           name={this.props.name}
-          id={this.props.name}
+          id={this.props.id}
           defaultValue={this.props.value}
           readOnly={false}
           defaultChecked={this.props.checked}